Second commit

+++ src/main/java/ru/netology/service/PostService.java
@@ -1,14 +1,18 @@
 package ru.netology.service;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 import ru.netology.model.Post;
 import ru.netology.repository.PostRepository;
 
 import java.util.List;
 import java.util.Optional;
 
+@Service
 public class PostService {
     private final PostRepository repository;
 
+    @Autowired
     public PostService(PostRepository repository) {
         this.repository = repository;
     }
@@ -26,6 +30,11 @@ public class PostService {
     }
 
     public boolean removeById(long id) {
-        return repository.removeById(id);
+        Optional<Post> post = repository.getById(id);
+        if (post.isPresent()) {
+            repository.removeById(id);
+            return true;
+        }
+        return false;
     }
 }
